Begin moving the play logic into its own class
Former-commit-id: c82c890f20eb7342ecdee77cfb4dd380b5bee377

+++ include/Menu.h
@@ -25,7 +25,7 @@ namespace SuperHaxagon {
 		int transitionFrame = 0;
 		int transitionDirection = 0;
 		int lastLevel = 0;
-		int level = 0;
+		size_t level = 0;
 	};
 }
